Added comments and send messages
