+++ src/constants.js
@@ -1,0 +1,4 @@
+export const DEFAULT_USER = {
+  id: 'chernandez',
+  name: 'Carlos Hernandez'
+};
